# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope:
    properties:
      code:
        example: "0000"
        type: string
      data: {}
      message:
        example: ok
        type: string
    type: object
...
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_v1_request.SubmitComment:
    properties:
      content:
        maxLength: 5000
        minLength: 1
        type: string
      parent_id:
        type: integer
    required:
    - content
    type: object
...
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_v1_response.CommentBasic:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      like:
        $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_v1_response.LikeInfo'
      parent_id:
        type: integer
      post_id:
        type: integer
      replies_count:
        type: integer
      user_id:
        type: integer
      user_profile:
        $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_v1_response.CommentUserProfile'
    type: object
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_v1_response.CommentUserProfile:
    properties:
      avatar:
        type: string
      bio:
        type: string
      blog_url:
        type: string
      email:
        type: string
      name:
        type: string
      region:
        type: string
      status:
        type: string
    type: object
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_v1_response.LikeInfo:
    properties:
      liked:
        type: boolean
      likes:
        type: integer
    type: object
...
info:
  contact: {}
  description: Xiaoqing Blog 后端 API（Admin + Public V1）。
  title: Xiaoqing Blog API
  version: "1.0"
paths:
...
  /v1/comments/{id}:
    delete:
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - BearerAuth: []
      summary: 删除自己的评论
      tags:
      - V1.Comments
  /v1/comments/{id}/likes:
    delete:
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_v1_response.LikeInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - BearerAuth: []
      summary: 取消点赞评论
      tags:
      - V1.Comments
    post:
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_v1_response.LikeInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - BearerAuth: []
      summary: 点赞/取消点赞评论
      tags:
      - V1.Comments
...
  /v1/content/posts/{id}/comments:
    get:
      description: 可选携带 BearerAuth，以返回点赞状态（like.liked）。
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_v1_response.CommentBasic'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      summary: 文章评论列表
      tags:
      - V1.Comments
    post:
      consumes:
      - application/json
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_v1_request.SubmitComment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - BearerAuth: []
      summary: 提交评论
      tags:
      - V1.Comments
...
schemes:
- http
- https
securityDefinitions:
  AdminSession:
    in: cookie
    name: fiber_session
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"